requirements.txt
flask
pandas
pylint
wikipedia
google-cloud
simplejson
